requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0

pydantic==2.7.1
pyyaml==6.0.1

faiss-cpu==1.8.0
rank-bm25==0.2.2

sentence-transformers==2.7.0
transformers==4.44.0
torch>=2.3.0


pymupdf==1.24.4
pytesseract==0.3.10        # optional if you install Tesseract OCR
opencv-python==4.10.0.84
Pillow==10.4.0

scikit-learn==1.5.1
numpy==1.26.4
pandas==2.2.2

regex==2024.7.24
python-multipart==0.0.9

torchvision==0.18.1
matplotlib==3.9.0
